Refactor console control and logging; remove unused functions, enhance argument parsing, and introduce console loop functionality
