+++ include/api/api.h
@@ -1,0 +1,16 @@
+//
+// Created by Jaguar Rivero on 2025-02-05.
+//
+
+#ifndef API_H
+#define API_H
+
+#include <iostream>
+#include <string>
+
+namespace api
+{
+    void execute(int &numProcesses, std::string &processType);
+}
+
+#endif //API_H
