+++ include/consolecontroller/console_control.h
@@ -3,10 +3,7 @@
 #include <string>
 namespace cli::driver
 {
-    void                      consoleLoop(bool run = true);
     void                      parseArguments(int argc, char *argv[], int &numProcesses, std::string &processType);
-    template<typename T> void printpid(const std::string &str, const T &x = T());
-    int                       LoggerExample();
 
 } // namespace cli::driver
