+++ src/ProcessController.cpp
@@ -1,10 +1,8 @@
-#include <iostream>
-#include <thread>
-#include "console_control.h"
-#include "logger_instance.h"
-#include "nap_time.h"
-#include "main_controller.h"
 
+#include "api.h"
+#include "console_control.h"
+#include "nap_time.h"
+#include <thread>
 void displayCompilationInfo(const char *appName)
 {
     std::cout << "*******************************************" << std::endl;
@@ -22,15 +20,9 @@ auto main(int argc, char *argv[]) -> int
 
     int         numProcesses = 4;
     std::string processType  = "simul";
-
     cli::driver::parseArguments(argc, argv, numProcesses, processType);
-
-    tools::LoggerManager::createLoggerType();
-
-    process::MainController::run(processType, numProcesses);
-
-    cli::driver::printpid("[PARENT PROCESS] Main process exiting", "");
-
+    api::execute(numProcesses, processType);
+    std::cout << "[PARENT PROCESS] Main process exiting" << std::endl;
     std::this_thread::sleep_for(std::chrono::milliseconds(tools::NapTimeMs::LONG));
 
     return 0;
