+++ src/consolecontroller/console_control.cpp
@@ -1,42 +1,19 @@
 #include "console_control.h"
-#include <atomic>
 #include <iostream>
 #include <thread>
 #include <unistd.h>
-#include "cli_controller.h"
 #include "console_logger.h"
 #include "logger_instance.h"
 #include "main_controller.h"
 #include "nap_time.h"
 #include "string_tools.h"
+#include "console_loop.h"
+
 namespace cli::driver
 {
-    void killPid(pid_t pid);
-    void terminatePid(pid_t pid);
-    void intPid(pid_t pid);
-    void doCommand(const std::string &input);
-    void printContext(int numProcesses = -1, const std::string &processType = "");
-    void printCommands();
+
     static tools::ConsoleLogger cl;
-
-    template<typename T> void printpid(const std::string &str, const T &x)
-    {
-        cl << str << " " << x;
-        cl.flush(tools::LogLevel::INFO);
-    }
-
-    template<typename T> void printpidE(const std::string &str, const T &x)
-    {
-
-        cl << str << " " << x;
-        cl.flush(tools::LogLevel::ERROR);
-    }
-
-    template<typename T> void printpidW(const std::string &str, const T &x)
-    {
-        cl << str << " " << x;
-        cl.flush(tools::LogLevel::WARNING);
-    }
+    static constexpr int MAX_PROCESSES = 20;
 
     void parseArguments(int argc, char *argv[], int &numProcesses, std::string &processType)
     {
@@ -50,11 +27,11 @@ namespace cli::driver
                     numProcesses = std::atoi(optarg);
                     if (numProcesses <= 0)
                     {
-                        printpidW("Number of processes must be greater than 0. Defaulting to ", 4);
+                        cl << "Number of processes must be greater than 0. Defaulting to 4.";
+                        cl.flush(tools::LogLevel::WARNING);
                         numProcesses = 4;
                     }
-                    numProcesses = std::min(numProcesses, 20);
-
+                    numProcesses = std::min(numProcesses, MAX_PROCESSES);
                     break;
                 case 't':
                     // Set the process type from the argument
@@ -62,7 +39,8 @@ namespace cli::driver
                     if (processType != "real" && processType != "simul" && processType != "network" &&
                         processType != "system")
                     {
-                        printpidW("Invalid process type defaulting to ", "simul");
+                        cl << "Valid process types are: real, network, system, simul. Defaulting to simul.";
+                        cl.flush(tools::LogLevel::WARNING);
                         processType = "simul";
                     }
                     break;
@@ -77,7 +55,8 @@ namespace cli::driver
                     }
                     else
                     {
-                        printpidW("Invalid logging type defaulting to ", "console");
+                        cl << "Invalid logging type defaulting to " << "console";
+                        cl.flush(tools::LogLevel::WARNING);
                         tools::LoggerManager::loggerType() = "console";
                     }
                     break;
@@ -97,20 +76,24 @@ namespace cli::driver
                     }
                     else
                     {
-                        printpidW("Invalid nap time type defaulting to ", "MS");
+                        cl << "Invalid nap time type defaulting to " << "MS";
+                        cl.flush(tools::LogLevel::WARNING);
                         tools::SleepTime::type = tools::NapType::MS;
                     }
                     break;
                 case 'h':
                 default:
                     // Display usage information and exit
-                    printpid(argv[0], "Usage:\n"
-                                      "-n <number of processes> (max 20)\n"
-                                      "-t <process type 'real', 'network', 'system' or 'simul' (default)>\n"
-                                      "-s <respawn (0 or 1)>\n"
-                                      "-l <logging type 'console' or 'file'>\n"
-                                      "-T <nap time type 'MS', 'SEC', 'MIN'>\n"
-                                      "-h -> help");
+                    cl  << argv[0]
+                        << "\n"
+                        << "Usage:\n"
+                        << "-n <number of processes> (max 20)\n"
+                        << "-t <process type 'real', 'network', 'system' or 'simul' (default)>\n"
+                        << "-s <respawn (0 or 1)>\n"
+                        << "-l <logging type 'console' or 'file'>\n"
+                        << "-T <nap time type 'MS', 'SEC', 'MIN'>\n"
+                        << "-h -> help";
+                    cl.flush(tools::LogLevel::INFO);
                     std::exit(EXIT_SUCCESS);
             }
         }
@@ -118,209 +101,10 @@ namespace cli::driver
         printCommands(); // Call to printHelp
     }
 
-    void printContext(int numProcesses, const std::string &processType)
-    {
-        // Static variables to store the latest values
-        static int         lastNumProcesses = 0;
-        static std::string lastProcessType  = "unknown";
-
-        // Update static variables only if new values are provided
-        if (numProcesses != -1)
-            lastNumProcesses = numProcesses;
-        if (!processType.empty())
-            lastProcessType = processType;
-
-        // Print the stored context
-        std::cout << "\n========================= Context =========================\n"
-                  << " PID                 : " << getpid() << "\n"
-                  << " Number of Processes : " << lastNumProcesses << "\n"
-                  << " Process Type        : " << lastProcessType << "\n"
-                  << " Respawn             : " << (process::ProcessController::respawn() ? "Enabled" : "Disabled") << "\n"
-                  << " Logging Type        : " << process::ProcessController::loggingTypeToString() << "\n"
-                  << " Nap Time Type       : " << tools::SleepTime::NapTypeToString() << "\n"
-                  << "==========================================================\n\n"
-                  << std::flush;
-    }
-
-    void printCommands()
-    {
-        std::cout
-                << "\n==========================================================\n"
-                << "Process Commands Help Menu (" << "PID: " << getpid() << ")\n"
-                << "==========================================================\n"
-                << "Available commands:\n"
-                << "  context         - Display the current context\n"
-                << "  quit            - Signals the program to gracefully quit at the next loop\n"
-                << "  term all        - Terminate (SIGTERM) all processes and exit the program\n"
-                << "  term <pid>      - Terminate (SIGTERM) a specific process by PID\n"
-                << "  int all         - Interrupt (SIGINT) all processes and exit the program\n"
-                << "  int <pid>       - Interrupt (SIGINT) a specific process by PID\n"
-                << "  kill all        - Kill all (SIGKILL) processes and exit the program\n"
-                << "  kill <pid>      - Kill a (SIGKILL) specific process by PID\n"
-                << "  display pids    - Display all current PIDs\n"
-                << "  respawn on      - Turn on respawn\n"
-                << "  respawn off     - Turn off respawn\n"
-                << "  monitor on      - Turn on monitoring: spawn monitoring threads\n"
-                << "  monitor off     - Turn off monitoring: end monitoring threads\n"
-                << "  help            - Display this help message\n\n"
-                << "==========================================================\n\n"
-                << std::flush;
-    }
-
-    void consoleLoop(bool run)
-    {
-         static std::unique_ptr<cli::driver::CLIController> cc{nullptr};
-        if (run)
-        {
-            if(!cc)
-            {
-                cc = std::make_unique<cli::driver::CLIController>();
-                cc->run(doCommand);
-            }
-        }
-        else
-        {
-            if(cc)
-            {
-                cc->stop();
-                cc.reset();
-            }
-        }
-    }
-
-    void doCommand(const std::string &input)
-    {
-        if (input.empty())
-        {
-            return;
-        }
-
-        if (input == "context")
-        {
-            printContext();
-        }
-        else if (input == "quit")
-        {
-            process::ProcessController::running() = false;
-            printpid("[QUIT] Signalling the program to gracefully quit at the next loop.", "");
-        }
-        else if (input == "term all")
-        {
-            printpid("[TERMINATE ALL] Terminating all processes and exiting.", "");
-            process::ProcessController::terminateAll();
-        }
-        else if (input.rfind("term ", 0) == 0)
-        {
-            auto vals = tools::string::splitString(input);
-            printpid("[TERMINATE ONE] Terminating process with PID:", vals[1]);
-            terminatePid(std::stoi(vals[1]));
-        }
-        else if (input == "int all")
-        {
-            printpid("[INTERRUPT ALL] Interrupting all processes and exiting.", "");
-            process::ProcessController::intAll();
-        }
-        else if (input.rfind("int ", 0) == 0)
-        {
-            auto vals = tools::string::splitString(input);
-
-            printpid("[INTERRUPT ONE] Interrupting process with PID:", vals[1]);
-            intPid( std::stoi(vals[1]) );
-        }
-        else if (input == "kill all")
-        {
-            printpid("[KILL ALL] Killing all processes and exiting.", "");
-            process::ProcessController::killAll();
-        }
-        else if (input == "monitor on")
-        {
-            process::ProcessController::continueMonitoring();
-        }
-        else if (input == "monitor off")
-        {
-            process::ProcessController::pauseMonitoring();
-        }
-        else if (input.rfind("kill ", 0) == 0)
-        {
-            auto vals = tools::string::splitString(input);
-            printpid("[KILL ONE] Killing process with PID:", vals[1]);
-            killPid(std::stoi(vals[1]));
-        }
-        else if (input == "display pids")
-        {
-            process::ProcessController::displayAllPids();
-        }
-        else if (input == "help")
-        {
-            printCommands();
-        }
-        else if (input == "respawn on")
-        {
-            process::ProcessController::respawn() = true;
-            printpid("[RESPAWN] Respawn feature is now", "ON");
-        }
-        else if (input == "respawn off")
-        {
-            process::ProcessController::respawn() = false;
-            printpid("[RESPAWN] Respawn feature is now", "OFF");
-        }
-        else
-        {
-            printpidE("[UNK] Unknown command [" + input + " ] Type 'help' for a list of available commands.", "");
-        }
-    }
-
-    void killPid(pid_t pid)
-    {
-        try
-        {
-            process::ProcessController::killProcessByPid(pid);
-        }
-        catch (const std::invalid_argument &)
-        {
-            printpidE("Invalid PID format.", "");
-        }
-        catch (const std::out_of_range &)
-        {
-            printpidE("PID out of range.", "");
-        }
-    }
-
-    void terminatePid(pid_t pid)
-    {
-        try
-        {
-            process::ProcessController::terminateProcessByPid(pid);
-        }
-        catch (const std::invalid_argument &)
-        {
-            printpidE("Invalid PID format.", "");
-        }
-        catch (const std::out_of_range &)
-        {
-            printpidE("PID out of range.", "");
-        }
-    }
-
-    void intPid(pid_t pid)
-    {
-        try
-        {
-            process::ProcessController::intProcessByPid(pid);
-        }
-        catch (const std::invalid_argument &)
-        {
-            printpidE("Invalid PID format.", "");
-        }
-        catch (const std::out_of_range &)
-        {
-            printpidE("PID out of range.", "");
-        }
-    }
-
     int LoggerExample()
     {
         //
+        tools::ConsoleLogger cl;
         cl.log(tools::LogLevel::INFO, "This is an info message.");
         cl.log(tools::LogLevel::WARNING, "This is a warning message.");
         cl.log(tools::LogLevel::ERROR, "This is an error message.");
